requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0

# UI
streamlit>=1.30.0

# Utilities
requests>=2.31.0